+++ cmsaf/inst/toolbox/server.R
@@ -3,7 +3,7 @@
 # You should not use this R-script on its own!
 #
 # Have fun with the CM SAF R TOOLBOX!
-#                                              (Steffen Kothe / CM SAF 2021-02-11)
+#                                              (Steffen Kothe / CM SAF 2021-08-17)
 #__________________________________________________________________________________
 
 # Function to compute first of month
@@ -320,6 +320,7 @@ function(input, output, session) {
   # Reactive values for coordinate bounds in visualize
   lon_bounds <- reactiveVal()
   lat_bounds <- reactiveVal()
+  ihsf       <- reactiveVal()
 
   # Longitute, latitute location vectors (of own locations added in options)
   lon_loc_vec  <- reactiveVal()
@@ -1673,7 +1674,7 @@ function(input, output, session) {
       })
     }
     
-    # Rendering lat, lon range.
+    # Rendering lon range.
     output$lonRange_ui <- renderUI({
       sliderInput(inputId = "lonRange",
                   label = "Please select a longitude range.",
@@ -1684,7 +1685,7 @@ function(input, output, session) {
       )
     })
 
-    # Rendering lat, lat range.
+    # Rendering lat range.
     output$latRange_ui <- renderUI({
       sliderInput(inputId = "latRange",
                   label = "Please select a latitude range.",
@@ -2435,21 +2436,55 @@ function(input, output, session) {
       lat <- file_data$dimension_data$y
       
       output$region_to_select <- renderUI({
+        # tags$div(id = "region",
+        #          sliderInput("lonRegionSlider",
+        #                      label = "Select longitude",
+        #                      min = ceiling(min(lon, na.rm = T)*100)/100,
+        #                      max = floor(max(lon, na.rm = T)*100)/100,
+        #                      value = c(ceiling(min(lon, na.rm = T)*100)/100,
+        #                                floor(max(lon, na.rm = T)*100)/100),
+        #                      width = "320px"),
+        #          sliderInput("latRegionSlider",
+        #                      label = "Select latitude",
+        #                      min = ceiling(min(lat, na.rm = T)*100)/100,
+        #                      max = floor(max(lat, na.rm = T)*100)/100,
+        #                      value = c(ceiling(min(lat, na.rm = T)*100)/100,
+        #                                floor(max(lat, na.rm = T)*100)/100),
+        #                      width = "320px"))
         tags$div(id = "region",
-                 sliderInput("lonRegionSlider",
-                             label = "Select longitude",
-                             min = ceiling(min(lon, na.rm = T)*100)/100,
-                             max = floor(max(lon, na.rm = T)*100)/100,
-                             value = c(ceiling(min(lon, na.rm = T)*100)/100,
-                                       floor(max(lon, na.rm = T)*100)/100),
-                             width = "320px"),
-                 sliderInput("latRegionSlider",
-                             label = "Select latitude",
-                             min = ceiling(min(lat, na.rm = T)*100)/100,
-                             max = floor(max(lat, na.rm = T)*100)/100,
-                             value = c(ceiling(min(lat, na.rm = T)*100)/100,
-                                       floor(max(lat, na.rm = T)*100)/100),
-                             width = "320px"))
+          fluidRow(
+            column(width = 5,
+                   numericInput("lonRegionMin",
+                               label = "Longitude min",
+                               min = ceiling(min(lon, na.rm = T)*100)/100,
+                               max = floor(max(lon, na.rm = T)*100)/100,
+                               value = ceiling(min(lon, na.rm = T)*100)/100)
+                               ),
+            column(width = 5,
+                   numericInput("lonRegionMax",
+                                label = "Longitude max",
+                                min = ceiling(min(lon, na.rm = T)*100)/100,
+                                max = floor(max(lon, na.rm = T)*100)/100,
+                                value = floor(max(lon, na.rm = T)*100)/100
+                   ))),
+            fluidRow(
+              column(width = 5,
+                     numericInput("latRegionMin",
+                                  label = "Latitude min",
+                                  min = ceiling(min(lat, na.rm = T)*100)/100,
+                                  max = floor(max(lat, na.rm = T)*100)/100,
+                                  value = ceiling(min(lat, na.rm = T)*100)/100)
+              ),
+              column(width = 5,
+                     numericInput("latRegionMax",
+                                  label = "Latitude max",
+                                  min = ceiling(min(lat, na.rm = T)*100)/100,
+                                  max = floor(max(lat, na.rm = T)*100)/100,
+                                  value = floor(max(lat, na.rm = T)*100)/100
+                     ))
+            
+          )
+        )
       })
     }
 
@@ -2801,10 +2836,14 @@ function(input, output, session) {
       argumentList <- list(var = input$usedVariable,
                            infile = nc_path_analyze(),
                            outfile = newOutfile,
-                           lon1 = input$lonRegionSlider[1],
-                           lon2 = input$lonRegionSlider[2],
-                           lat1 = input$latRegionSlider[1],
-                           lat2 = input$latRegionSlider[2],
+                           # lon1 = input$lonRegionSlider[1],
+                           # lon2 = input$lonRegionSlider[2],
+                           # lat1 = input$latRegionSlider[1],
+                           # lat2 = input$latRegionSlider[2],
+                           lon1 = input$lonRegionMin,
+                           lon2 = input$lonRegionMax,
+                           lat1 = input$latRegionMin,
+                           lat2 = input$latRegionMax,
                            nc34 = input$format,
                            overwrite = TRUE)
     } else if (currentOperatorOption() == "point") {
@@ -3454,10 +3493,14 @@ function(input, output, session) {
         start_date = input$dateRange_analyze[1],
         end_date = input$dateRange_analyze[2],
         country_code = input$country,
-        lon_min = input$lonRegionSlider[1],
-        lon_max = input$lonRegionSlider[2],
-        lat_min = input$latRegionSlider[1],
-        lat_max = input$latRegionSlider[2],
+        # lon_min = input$lonRegionSlider[1],
+        # lon_max = input$lonRegionSlider[2],
+        # lat_min = input$latRegionSlider[1],
+        # lat_max = input$latRegionSlider[2],
+        lon1 = input$lonRegionMin,
+        lon2 = input$lonRegionMax,
+        lat1 = input$latRegionMin,
+        lat2 = input$latRegionMax,
         out_dir = monitor_climate_out_dir,
         temp_dir = monitor_climate_temp_dir,
         climate_dir = monitor_climate_temp_dir,
@@ -3572,8 +3615,10 @@ function(input, output, session) {
         } else if (currentOperatorOption() == "point") {
           newRow <- data.frame(operatorInput_value(), "point", paste0("lat: ", input$latPoint, ", lon: ", input$lonPoint))
         } else if (currentOperatorOption() == "region") {
-          newRow <- data.frame(operatorInput_value(), "region", paste0("lat: [", input$latRegionSlider[1], " ", input$latRegionSlider[2], "], ",
-                                                                     "lon: [", input$lonRegionSlider[1], " ", input$lonRegionSlider[2], "]"))
+          # newRow <- data.frame(operatorInput_value(), "region", paste0("lat: [", input$latRegionSlider[1], " ", input$latRegionSlider[2], "], ",
+          #                                                            "lon: [", input$lonRegionSlider[1], " ", input$lonRegionSlider[2], "]"))
+          newRow <- data.frame(operatorInput_value(), "region", paste0("lat: [", input$latRegionMin, " ", input$latRegionMax, "], ",
+                                                                       "lon: [", input$lonRegionMin, " ", input$lonRegionMax, "]"))
         } else if (currentOperatorOption() == "dateRange") {
           newRow <- data.frame(operatorInput_value(), "dateRange", paste0("from ", input$dateRange_analyze[1], " to ", input$dateRange_analyze[1]))
         } else {
@@ -4456,6 +4501,15 @@ function(input, output, session) {
                         value = c(tmp[1], tmp[2]))
           })
           
+          output$ihsf_visualize <- renderUI({
+            tmp <- 0.1
+            ihsf(tmp)
+            sliderInput("decimal",
+                        label = "Image Ratio",
+                        min = -0.9, max = 0.9,
+                        value = 0.1, ticks = FALSE)
+          })
+          
           output$title_text <- renderUI({
             textInput("text1",
                       label = "Title",
@@ -4952,6 +5006,7 @@ function(input, output, session) {
   observeEvent({
     lon_bounds()
     lat_bounds()
+    ihsf()
   }, {
 
     req(visualizeVariables()$plot_dim == 2)
@@ -4961,7 +5016,7 @@ function(input, output, session) {
       lon_bounds = lon_bounds(),
       lat_bounds = lat_bounds(),
       image_def = image_def,
-      ihsf = ihsf
+      ihsf = ihsf()
       )
 
     imagewidth(imDim$imagewidth)
@@ -5933,7 +5988,7 @@ function(input, output, session) {
                                                       plot_grid = plot_grid,
                                                       grid_col = grid_col,
                                                       image_def = image_def,
-                                                      ihsf = ihsf))
+                                                      ihsf = ihsf()))
       } else {
         if(checkboxCompareData_dropdown() == "cmsaf.side.by.side"){
           req((visualizeVariables()$plot_type == "cmsaf.side.by.side"))
@@ -6094,6 +6149,7 @@ function(input, output, session) {
     if (is.null(input$zoom_brush)) {
       lon_bounds(input$slider1)
       lat_bounds(input$slider2)
+      ihsf(input$decimal)
     } else {
       brush <- input$zoom_brush
       lon <- c(brush$xmin, brush$xmax)
@@ -6174,7 +6230,7 @@ function(input, output, session) {
                                                                   plot_grid = plot_grid,
                                                                   grid_col = grid_col,
                                                                   image_def = image_def,
-                                                                  ihsf = ihsf)
+                                                                  ihsf = ihsf())
                          in_plot <- res_plot$src
                        } else {
                          res_plot <- cmsafvis::render_plot(plot_rinstat = input$plot_rinstat,